Explicitly support proxying storage through server
