+++ packages/server/src/fhir/rewrite.ts
@@ -1,7 +1,7 @@
 import { Binary, Resource } from '@medplum/fhirtypes';
 import { getConfig } from '../config/loader';
 import { getLogger } from '../logger';
-import { getBinaryStorage } from '../storage/loader';
+import { getPresignedUrl } from '../storage/loader';
 import { Repository } from './repo';
 
 /**
@@ -168,7 +168,7 @@ class Rewriter {
       return `Binary/${id}`;
     }
 
-    return getBinaryStorage().getPresignedUrl(binary);
+    return getPresignedUrl(binary);
   }
 }
